Make NetcdfDataSource a wrapper for results.
The NetcdfDataSource class basically contains all the components that
make up a 3Di result and are needed to extract all the relevant data.
These are: the netcdf itself (subgrid_map.nc), id_mapping.json, and
the aggregated netcdf flow_aggragate.nc. The docstrings have been
updated to reflect this change.

Additionally, a bunch of deprecated stuff was removed.