Add proper deactivation of route tool.
The bug was that if you click on the "kies sideview traject" button and
then click on another tool (e.g. zoom tool) and then again click on
"kies sideview traject" button it wouldnt switch to the tool. You have
to click the button again for it to switch to the route tool.