Add chunked downloading, thread cleanup, improve Endpoint initialization
Downloading was processed entirely in memory per file, which blows up
memory usage for large files, thus we needed to download in chunks and
process the chunks using the `readyRead` signal.

The thread that connects to the Lizard API wasn't correctly closed, so a
`stop` method was added to gracefully stop the thread.

Slight detail: lizard-connector Endpoint should be initialized using a
single get before we can use it as an iterator.

Some minor UX improvements.