Gracefully handle HTTPError when trying to get ThreediResults.
ThreeDiResultSelectionWidget starts a new thread to fetch results
from lizard in the background as soon as the user clicks on 'login'.
However, any errors thrown when connecting to lizard (most notably
authorization errors) are not handled gracefully which cause the
whole qgis application to crash.

I catch the HTTPError and added an extra pyqtSignal `connection_failure`,
which is called when the error occurs. This signals the main app of the
error and provides the status and reason.