Change fix for content_type = v2_added_c
The previous fix was done in the styling, but some tools (e.g. Animation Toolbar) look at the 'type' field to determine wether or not to reverse the flow direction. Positive flow in 1D-2D links is opposite to the drawing direction (known issue).

This new fix sets the 'type' attribute to '1d_2d', so that all tools regard it in the same way as any other type of 1D-2D flowline